add lombok annotation
Took 2 minutes

+++ src/main/java/at/technikum/DAL/CardData.java
@@ -16,7 +16,7 @@ public class CardData {
     }
 
     // GET /weather/:id
-    public Card getCard(Integer ID) {
+    public Card getCard(Long ID) {
         Card foundCard = cardData.stream()
                 .filter(card -> ID == card.getId())
                 .findAny()
